Remove AbstractRepository.
