+++ MANO-API/src/main/java/com/ubiqube/etsi/mano/repository/msa/SubscriptionMsa.java
@@ -1,8 +1,5 @@
 package com.ubiqube.etsi.mano.repository.msa;
 
-import java.io.File;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.UUID;
 
 import org.slf4j.Logger;
@@ -10,10 +7,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.ubiqube.api.exception.ServiceException;
 import com.ubiqube.api.interfaces.repository.RepositoryService;
-import com.ubiqube.etsi.mano.exception.GenericException;
-import com.ubiqube.etsi.mano.grammar.AstBuilder;
 import com.ubiqube.etsi.mano.grammar.JsonFilter;
 import com.ubiqube.etsi.mano.model.vnf.sol005.SubscriptionObject;
 import com.ubiqube.etsi.mano.repository.SubscriptionRepository;
@@ -61,25 +55,4 @@ public class SubscriptionMsa extends AbstractGenericRepository<SubscriptionObjec
 		return "susbscription.json";
 	}
 
-	@Override
-	public List<SubscriptionObject> query(final String _query) {
-		final AstBuilder astBuilder = new AstBuilder(_query);
-		List<String> listFilesInFolder;
-		try {
-			listFilesInFolder = repositoryService.doSearch(REPOSITORY_SUBSCRIPTION_BASE_PATH, "");
-		} catch (final ServiceException e) {
-			throw new GenericException(e);
-		}
-		final ArrayList<SubscriptionObject> ret = new ArrayList<>();
-		for (final String entry : listFilesInFolder) {
-			final String path = entry.substring((REPOSITORY_SUBSCRIPTION_BASE_PATH + '/').length());
-			final File file = new File(path);
-			LOG.info("Retreiving: {}", file.getParent());
-			final SubscriptionObject subscriptionObject = get(file.getParent());
-			if (jsonFilter.apply(subscriptionObject.getSubscriptionsPkgmSubscription(), astBuilder)) {
-				ret.add(subscriptionObject);
-			}
-		}
-		return ret;
-	}
 }
